Unify and clean logging usage across tests
To help a bit humans when reading the tests logs.

It seems like there was a design for reusing the logger (i.e. having the
test name in front of each log line) but has been lost when expanding
the framework and adding tests.

This commit tries to unify and clean the usage:
- Remove the completely unused logger in tests/__init__.py
- Force all tests to use the MenderAPI logger object
- Make docker_compose_manager use the root logger (that, surprise!, is
the one containing the test name)
- Substitute the log hook in pytest_runtest_setup by a fixture; note
that to support having installed or not xdist the def is rather hacky.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>