dump-gitlab-env: pull and display gitlab envvars
some of the env is out there in compose files, but
some is hidden in gitlab (mender-qa).

to get reliable local test runs it's good to have
them pulled from the single source of truth.

this util pulls the env vars from the gitlab project and
dumps them in a format suitable for sourcing to your shell.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>