Reuse docker-compose environments between tests
This commit modifies the scope of the fixtures setting up the
docker-compose environment between test runs. Setting the scope
to session makes the tests reuse the environment between runs.

In order to avoid overlapping setups, which is likely to break
badly the tests, we run pytest multiple times selecting each
time one of the marker to run the tests which share the common
environment.

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>