release_tool: Work around error when tracking local integration/master.
This is typically the checked out branch, so Git will complain if you
try to fetch into it. However, we don't care about that since this is
a temporary repository clone. So just check out a temporary commit and
fetch into master which is now not checked out anymore.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>