QA-170: Move backend-tests files knowledge to integration repo
This commit introduces a major rework in the ./run script so that the
knowledge of which fails goes with which suite (open/enterprise) is in
this repo rather than in the caller (i.e. mender-qa, which is not
versioned).

The motivation is to avoid once the recurrent problem of breaking old
releases when introducing/removing files

Now the wrapper takes a -s|--suite parameter that can be open (default),
enterprise or all and it takes care of the files required in each case.
The -f parameter to specify files is now deprecated

As a side effect, we have also moved the results pytest args here, so
that we can control that the "all" case does not override results.

It still accept user custom filters or other pytest parameters.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>