Move get_valid_image to a session scoped fixture
And create a similar one for the rofs image. The point is to avoid using
the blobally scoped fabric.env following the design from client
acceptance tests.

It would have been nice to also remove the global machine_name creating
a similar fixture, but the problem is its usage from deep in the stack
in the helper methods common_update_procedure and similar.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>