QA-132: Migrage integration tests suite to Python 3
This commit brings all the necessary fixes to run the tests in Python 3.

The main work is around Fabric 2 upgrade: rewrite of the private methods
of MenderDevice and MenderDeviceGroup and clean-up of conftest.py.

Note that modifying MenderDeviceGroup to make use of the threading
execution has proved quite challenging and has been post-poned.

Other than that:
- pip dependencies updated
- run.sh wrapper updated to use python3
- Documentation updated
- And the actual Python3 code fixes which are mostly the byte-array to
string conversions.
- Modify logic in some specific tests that were expecting the old Fabric
stdout/Response object.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>