[integration staging] Tweak files to run integration tests Enterprise
These components will not have an staging branch, as they are not part
of our Hosted Mender setup.

On the one hand, replace the image for the OS repos with the Enterprise
ones, so that we run all integration tests with Enterprise binaries.

On the other hand, set the OS components as non-releasable for the
Pipeline not to build them.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>