[integration tests] Cleanup update_image and related tests
The main change is removing the wrongly used and misleading parameter
"skip_reboot_verification", which was only skipping a second check for
the device not reboot (?). The original intention of this is unknown.

Reorganized the parameters in update_image/update_image_fail to show
which ones are used by these and which are just passed downstream to
other helpers.

Simplified TestMultiTenancyEnterprise and optimize TestFaultTolerance.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>