Move dependency between deployments and storage-proxy to demo yaml
The docker-compose service mender-deployments depends on storage-proxy,
which is defined in docker-compose.demo.yml. This causes docker-compose
commands, as well as the logging command executed when the user presses
ENTER after running ./demo up, to fail with the following message:

ERROR: Service 'mender-deployments' depends on service 'storage-proxy'
which is undefined.

This commit moves the dependency on storage-proxy to the file where it
really belong to: docker-compose.demo.yml

Changelog: None

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>