[release_tool] MEN-4366: Use Mender version for Docker images on release
A new independent_component flag has been introduced to components-maps
which will mark the repos that will not follow the Mender product
version on their associated Docker images (currently only affects mender
client images).

Then, we use this information in different places of the tool to
differentiate which file to update (Git versions in git-versions, Docker
versions in docker-compose files) during the various parts of the
release workflow.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>