[release_tool] MEN-4366: Make --set-verion-of modify both versions
This option is used by CI to set the Docker verions to "pr" for local
testing, or by users to specify custom versions in feature branches. For
the second use case, setting both Docker and Git will be closer to what
the user expects.

Fix set_docker_compose_version_to to set all assosiated git repos when
the "repo" is actually a Component of type docker_container or
docker_image.

This way we also make --set-version-of and --version-of set and return
the same (Git versions).

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>