[release_tool] Make -f option run regardless of the directory name
The previous implementation would not work on an integration repo
checked out in a directory named something else.

The rest of the uses of find_upstream_remote are linked to the release
process or to --select-test-suite, and in both cases all repositories are
required (hence is okey to require the same directory name).

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>