[release_tool] Redesign for the new versioning schema (MEN-3466)
This commit modifies the original design following last suggestions from
the review. The most important change is that now git-versions.yml (and
the new -enterprise counterpart) uses containers names as keys, and
therefore the exact same parsing principles that the rest.

Several parts of the tool have been affected, as now the code needs to
decided whether git or docker versions are desired before parsing the
yml files.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>