release_tool: Fix remote-name being dropped from version ranges.
This happens when using -g and -i together with a range where one end
uses a remote reference and the other one doesn't. We are checking
whether the supplied reference is a remote or not, but only for the
first element of a range. We need to check it for all elements in the
range.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>