release_tool: Better recovery when tag deletion fails.
Do the local tag deletion after the remote deletion, instead of vice
versa. This means that we can immediately pick up where we left off if
we get an exception for whatever reason (network error, etc.). Without
this we have to refetch the remote tags in order to delete them
afterwards.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>