MEN-3606: Implement `./demo --enterprise-testing` flag.
This required some significant refactoring because with Enterprise we
have to launch the client in a separate step, instead of together as
in Open Source. Took the opportunity to also clean up a little and
standardize on number based truth values (both numbers and strings
were used before).

This is for internal testing and is not meant for consumption by
users. The client launching is not fully aligned with the onboarding
experience, because it instructs you to launch the client in a
separate step, whereas the demo script can only launch the backend and
the client in the same step (though inside the script it is two steps,
as described above).

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>