tests: Work around errors and warnings printed on stdout, not stderr.
The "payload" in the text is the Artifact name, but this is mixed with
warnings about not being able to connect to the syslog. This is not a
problem on later branches because stdout and stderr are properly
separated there.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit f3437167b6bb342c5bbb51220d441db44077f1f7)

Backported-by: Lluis Campos <lluis.campos@northern.tech>