Test binding to port 0
Fix the race condition which was previously present, in which the fact that
pytest-xdist runs parallell jobs in seperate processes, tricked up the global
port used in the RebootDetector class, since every process would start
allocating ports from the same number.

This fix make the allocation dependent upon the OS allocating an open port. This
is done when binding to port 0, and hence, this implementation should be both
process and thread safe. Plus, if you don't mind me saying. A fair bit more elegant.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>