Update deployment test to pass on staging
Introduce a multiplier for all the time offsets (start_ts, waiting time,
delays) to compensate the higher latency caused by the usage of a remote
application instead of a local docker-compose. These tests manipulate
+100 devices per time, and a 500ms per-call latency causes >50 seconds
to loop over all the devices, which makes the default time window size
(15 seconds) insufficient.

Set the multiplier to 8 if running on Kubernetes, which makes the
default time window size equal to 120 seconds.

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>