test: Fix race condition in test_compatibility
The race condition was present, since the clients ran in parallel with the
pytest, and since the clients were so quick to reboot that the reboot-detectors
did not have time to initialize fully.

This fix adds a barrier to a ParallelRebootDetector which has a guard which
waits for the reboot-detectors to come up, before the code within the 'with'
statement is released.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>