MEN-3075: Integration test for `clears_artifact_provides`.
We just need to modify and add some checks for the existing inventory
test to use the `clears_artifact_provides` defaults.

Note that deleting attributes does not work yet due to MEN-4001 and
MEN-4003, so the test is marked XFAIL at the moment. The rest appears
to work correctly.

Changelog: None
Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>