FIX test_devauth, testutils: after service restart, wait for traefik
dynamic configuration means we have to wait a bit for it to
take effect.

in this case, after killing the service, traefik uninstalls
the associated router and service. even when the service is back up,
traefik might not have yet noticed, and fail/reroute the requests, so:

- wait a bit for the restarted router/service to be registered again.

very sneaky errors can be experienced otherwise:
 - in this case, POST /auth_requests would go to mender-gui(!),
because of its catchall regex middlewares.
- resulted in: 405(!) from node's nginx (POST on a static resource...)

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>