FIX compose: fix gateway dependency on deployments/minio
gateway cannot depend on deployments, because it's actually
deployments that need the gateway exposing the s3.docker.mender.io
alias to work. that's because deployments pings minio on startup.

remove the depends_on, but do introduce a health check for deployments - and only register
the router if healthy. this will prevent pinging deployments when it's
unstable (waiting on minio, restarting, etc.).

also, fix the gateway depends_on minio - move it to minio layer, and
make sure the service is healthy first. this dep is not strictly necessary,
but nothing will happen without minio anyway.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>