FIX compose: fix gateway dependency on deployments
gateway cannot depends_on deployments, because it's actually
deployments that need the gateway exposing the s3.docker.mender.io
alias to work. that's because deployments pings minio on startup:

- relax the depends_on
- but do introduce a health check for deployments - and only register
the router if healthy. this will prevent pinging deployments when it's
unstable (waiting on minio, restarting, etc.). treat it as a replacement
for depends_on.

as a hint, also make the gateway depend on minio instead (if healthy).
without this nothing will happen anyway.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>