FIX: backend-tests: wait until traefik ready
first off - a helpful depends_on. this doesn't
guarantee too much, but it helps not starting the tests
way too early, before the gateway has any chance of
responding.

then - wait for all routers to be registered in traefik.
this is the closest guarantee we can get regarding traefik
'availability'.

(as a side effect, this also gives us a lot of time:
- traefik will register deployments only if it's healthy
- deployments will first wait for minio (~30secs) to be healthy)

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>