FIX test_devauth assertion checking
this started failing with:

E       AssertionError: assert 'assert 0 == 1' in ''
E        +  where '' = str(AssertionError())
E        +    where AssertionError() = <ExceptionInfo AssertionError()
tblen=4>.value

not sure why it worked at all - an assert that's not decorated
with a message will not have the message. and this is what we're
waiting for, a plain 'assert len(found) == 1'...

unclear what happened here, but adding proper annotation to the
assert we're expecting. indeed the logic itself works, and we
get what we want, now with a message that leaves no doubts.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>