demo script: Fix misparsing of commands when using `-f` argument.
If the `-f` argument is specified, we jump to the next parsing loop,
which doesn't contain parsing for docker commands. Move the parsing
there so it will be done correctly. This is fine because the first
loop will exit on unknown arguments, so we are guaranteed to get to
the second loop.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>