external storage-proxy support
almost the same storage proxy is reintroduced for legacy clients, but
this time it's 'external' - it forwards requests to the main gateway
and further to minio.

while the normal minio routing happens based on Host header,
the request from external proxy is recognized via the X-Forwarded-Host
header. it's set by the proxy ootb and is a standard header.

this way, traefik can accept http/2 ready clients directly, and http1.1
clients via the proxy almost transparently.

also, the s3.docker.mender.io domain is un-hardcoded from the minio
layer. it's strictly a demo value, and the rule has to be usable also in
prod - parametrize with an env var.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>