MEN-4088: Add integration tests for mender-shell.
In addition to the test, this includes adding:

* A protomsg library to encode Proto messages.
* A proto_shell library which speaks the mender-shell WebSocket
  protocol.
* A Websocket class which wraps the asynchronous Python websockets
  framework in a synchronous API. Much more convenient for tests that
  need to run strictly serially.

This also adds deviceconnect as a standard service in the integration
tests, which is not the same as production. This should be ok though,
since deviceconnect is just sitting idle if not used, and will not
affect the other tests. It's not worth launching it conditionally.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>