tests/run.sh: Fix inability to pass filenames into run.sh.
Because the last argument is in quotes, it always gets included, even
if empty, and acts as a silent path wildcard which always includes all
files. Fix it by using ":+" notation to make the quoting conditional.
If the variable is empty it will not be included as an argument.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>