integration: actually wait for traefik
the wiat_for_traefik util was already in testutils, and meant to
be shared between integration and backend-integration.

by mistake however, it's only used in backend-integration.

work it into integration tests by piggybacking it on
wait_for_containers. this way each setup waits for its
number of containers, but instead of sleeping a random
60 sec, it actually waits for traefik to be set up.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>