integration: actually wait for traefik
this is what I promised - but forgot about it.
the wait_for_traefik util was already in testutils, and meant to
be shared between integration and backend-integration.

by mistake however, it's only used in backend-integration -
so add it to integration as well.

work it into integration:
- add to DockerComposeNamespace._wait_for_containers method. this way each setup waits for
the gateway and basic routes as well
- don't sleep a random 60 sec anymore - if we have traefik and num of
containers, we should be good

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>