test_demo_artifact: stability fixes
- manage testprefix on manual restarts
- manually wait for traefik before doing anything

our manual restarts don't manage these two aspects.

the testprefix is lost (infra sets it for the 'up' cmd - but we don't).
results are unpredictable here because setups get mixed up with xdist.

also, only the infra's 'up' command waits for traefik - it's not the case for the
many restarts we do here. it's easy to be too early, as in some cases
even sleep(60) is not enough.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>