tests/test_configuration: detect update race condition
it can occur that ordinary update poll happens just
when a forced update-check is requested.

the update poll mechanism then wins and starts the update.
if the timing is close enough, the client will actively refuse
the forced check, and log the extra line we're
detecting here. we won't get the 'Forced wake-up...' part though.

if we hit this race condition, simply retry. if a couple
tries in a row fail, we might have a real problem.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>