Flaky PoC
This is just a test for verifying the use of the flaky package in order to rerun
tests which fail.

The PoC is added only to 'test_demo_artifact' to begin with, as this is a common
culprit of pipeline failures.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
(cherry picked from commit f22ac5eaf60c2d9df575eeea38458962ffb52216)