[release tools] MEN-4513: Changelog and statistics for Mender server
From now on we need to group the "Mender server" for changelog and
statistics, leaving everything else as independent repos.

`release_tool` supports now a new flag `--only-backend` that, when used
in combination with `--list` skips the independent components. The other
tools `changelog-generator` and `statistics-generator` are updated to
use this tool (note that they need to add `integration` repo).

Added unit tests to verify backwards compatibility.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>