[backend-tests/run] Do not override exit code when running twice
The previous code, when running two tests suites, would always exit with
the retcode of the second one - potentially masking an error in the
first test suite.

Indirectly, this fixes an issue in staging branch, where skipping the OS
test suite would make the script exit with code 1.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>