[tests] Add client version fixture and mark File Transfer limits tests
The fixture uses release_tool to figure out which mender client is
being used. Tests marked with min_version higher that the one being used
will be skip.

This fixture is specifically designed for staging branch, for which we
use the "latest stable" mender client and some tests might be at times
incompatible. For the rest of the branches it won't really have any
effect, as the client version will be a branch and validate as more
recent that the required minimum version.

Mark File Transfer limits tests to be skip on current client version.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>