Fix spurious failure in test_rootfs_image_rejected (and maybe others).
The test fails if the client is too quick, because it looks for the
"inprogress" deployment status. However, by the time it gets to do
that check, the client may already be past that and the deployment is
"finished".

What we really want to look for is whether the deployment has left the
"pending" status, which is before any clients have picked it up.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>