test_image_download_retry_timeout: Rework update start detection
Previous code relied on a single call to MenderDevice.run, with retries
the command until it succeeds using an exponential back-off time between
tries. This test is test time sensitive, it needs to break the
connectivity after it starts the download but before it completes, and
sometimes it is missed between different SSH connection attempts.

The new code basically tries with a while loop in a single SSH
connection, which should be much more accurate.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>