statistics-generator: Skip missing repositories.
This makes it possible to run the generator on an old release, but
using a newer branch of integration. Normally this isn't necessary,
but it helps with the transition from mender-docs to
mender-docs-changelogs. We need to use a more recent branch of
integration in order to omit client components from the changelog, but
still need to be able to include those components in the legacy
changelog, by using the old branch of integration.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>