smtp.mock.yml: Correctly set network and not expose port
The port does not need to be exposed out of the Docker network. The
problem was a typo in the `networks` key for `local-smtp` which made the
container not connected to the Docker network. This was hidden by the
fact that the port was exposed outside, and the tests used this outside
interface.

This fix allows for multiple tests using smtp mock to run in parallel.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>