release_tool: Update extra files also from CLI call
This is an amend to commit c68e632, which correctly sets the versions of
the extra docker compose files when releasing, but missed them when
using the --set-version-of CLI argument.

Fixes test failures from 3.1.x release branch, where these files are set
to a non-existing version (iow not set to "pr" by CI script).

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>