QA-187: add support to run full integration tests in the staging env
This commit updates the full integration tests to make it possible to
run them against the staging environment running in Kubernetes. The
Kubernetes container manager extends the basic Docker Compose container
manager, because we run the clients using docker-compose, but doesn't
start the Mender backend in docker compose. Instead, it relies on AWS
EKS, a public end-point for the management and device APIs and kubectl
to port-forward the internal APIs.

This commit switcheds the providees/depends integration test to the
Mender Docker client with addons, built by this repository, instead of
using the Docker client from the `mender` reepository.

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>