QA-187: fix instability of backend-test test_auditlogs.py
Audit logs are sorted by date/time, but the format doesn't include
milliseconds if they are zero. For this reason, when comparing
'2021-09-21T04:31:34Z' and '2021-09-21T04:31:34.443Z', the sort result
is wrong and this causes a failure in the test.

To make it more robust, parse the string and use datetime objects.

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>