Fix incorrectly passing test_update_failover_server.
It was passing because no check was done to check that the second
server was actually working. It wasn't, so the client naturally fell
back to the other one.

With this commit we update the failover server to the new traefik
configuration files, which makes it work again. We are still not
checking that the server works, but this will come in the next commit
together with actual migration to the new server.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>