MEN-5216: test_monitor_client: Restart mender-client after network loss
In the context of MEN-5216, the actual API calls to the Mender servers
are done with the local proxy via Mender client.

Restart the service so that the client needs to resolve again
docker.mender.io and then gets the connection refused, correctly
simulating the offline scenarios of these tests.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>