tests: Introduce helper new_tenant_client between tests and env managers
The story behind this is that we have a hook in conftest to inspect test
environment in search for MenderDevice/MenderDeviceGroup instances. When
found, then we can print systemd logs from the Mender client(s).

This works fine for all OS setups, as common_setup fixtures uses this
convention of attaching the device to the environment. But on Enterprise
setups this is test specific, so it was forgotten most of the times.

Now this common.py level helper sits in the middle of the tests and the
env managers, so that it can create the Mender client container and
attach the MenderDevice instance to the test env (so that, yey, the logs
can be printed).

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>