QA-333: enable email-dependant integration tests running in staging
Enable email-dependant integration tests when running in staging for
both backend and full integration tests. It depends on the GMAIL_ADDRESS
and GMAIL_PASSWORD env variables, which point to a dedicated gmail
mailbox which will receive the emails from the staging environment.

For this reason, the users created by these tests are in the from:

  ci.email.tests+UUIDv4@mender.io

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>