changelog-generator: Fix rendering of Dependabot changelogs.
The first problem is that Dependabot encapsulates some of the
information inside "---" and "...", which renders really badly
(interpreted as headers). Replace that with a "```" code block, so
that we also get the line breaks to render correctly.

Second, Grav does not follow Markdown very well, at least not compared
to other Markdown parsers, and getting both the block mentioned above,
as well as bullet points, to render correctly, is tricky. We need to
do a couple of things:

1. Grav does not render bullet points correctly to more than two
   levels, unless four spaces of indentation are used. Although [1]
   does list four spaces as the correct number of spaces to use, it is
   extremely common to support two spaces, especially since this lines
   up with text much better than four spaces. But Grav's the boss, so
   make sure we use four spaces in our aggregated Dependabot changelog
   entries.

2. Code blocks are completely messed up if they appear in a list
   without a blank line separating it from the rest of the text in the
   given bullet point. There actually is a blank line in the
   Dependabot output already, but we are removing blank lines, so to
   be consistent we stop doing that everywhere. Previous comments
   indicated that this looks bad when rendered in a list because of
   the additional whitespace it introduces, but I've checked and it
   doesn't seem all that bad, even makes it more readable in some
   places, so I think we can do this.

[1] https://www.markdownguide.org/basic-syntax/

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit 6780b49c34b77d9b32bfbdcb4f33e1fa759cb05a)