release_tool: Fix incorrect results for many queries after N-to-N mapping.
We can no longer trust that docker_image names are uniquely bound to
a repository, since some images contain multiple repositories' code.
Therefore switch to looking up by Git name first, when checking
versions.

The requires some extra fixes though. As can be seen in the commit,
two entries for mender and monitor-client were missing, but this was
not evident until now because we looked up by docker_image, not git.
Adding these is trivial, but it means that older versions cannot be
queried correctly because they are missing these entries, and they
can't be fixed either because they are in public tags. So we
unfortunately need some special code to handle the legacy cases.

Changelog: none

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>