release_tool: Deprecate and remove yml_components() function.
This was meant as a goto function when you just want something unique
to refer to a component by. However, it is no longer possible to refer
uniquely to any component that you don't know the type of, so every
section needs to take the type into account.

This fixes confusion in the tool when querying different combinations
of component types and version types.

This changes some key aspects of the querying. It is no longer
possible to query combinations of components and output types that
would produce more than one answer.

Changelog: None
Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit b77d849726d950856974b29135af9831ff3c4d72)