Disable test which cannot pass until the PR is merged.
This test is subject to the same incorrect data as mentioned in the
commit titled:

  "release_tool: Fix incorrect results for many queries after N-to-N
  mapping."

However, we cannot fix it in this PR, because it is *master* that
needs to have the data, and it will not have it until this is merged.
So disable the test for now, and it will be immediately re-enabled
afterwards.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>