test_release_tool: Remove outdated pre-modification of yml files.
The removed section is supposed to make it easier to test on
non-master branches, but it has gotten out of date, and now contains
stale data, making some of the queries wrong. It also complicates
things by making the test run on different data than what is available
on the command line.

Instead of trying to keep this up to date, remove it, and make this
one test dependent on being executed on master, skipping it
otherwise. This makes the test more realistic, at the cost of losing
some coverage on the versioned branches.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit 6b01ee6d0e5210eb8e1341dca5718c3832af6b85)