Fix client components being listed as backend components.
For this we need to set client components back to independent
components.

The motivation for setting them to dependent components is to have the
client docker images be tagged with the integration version. However,
doing this has a lot of side effects, as the components are now
treated as backend components, and will be listed by `--only-backend`
for example.

This was attempted solved with the `core_component` method in previous
commits, but here we solve it in a different way: Allow git and
docker_image components to have different independent status. This way
we can set "mender" to independent, while "mender-client-qemu" is
dependent.

This also requires splitting up `set_docker_compose_version_to`. We
change it so that it only sets one component and its associated type
at once, instead of looping over all of them. It is the caller's
responsibility to loop over whatever components it needs, be it git or
docker_image.

We also make `--set-version-of` respect `--version-type`, which isn't
strictly necessary, but complements the functionality, and makes it
easier and more predictable to use. The existing, slightly wrong,
behavior of setting both git and docker_image versions simultaneously
is kept for backwards compatibility, but can be changed on the command
line with the `--version-type` argument.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit 131d1efebd2a6a7a2a2465ee06937ed8c775d1fa)