Add meta-mender to components.
The motivation behind this is not to do releases of meta-mender using
release_tool; for this the process is too different. It is to have
some place to track the current latest branch of meta-mender, which
will be used by the changelog generator.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit f381795b4429b5f60a257929811d1318bf8e9d5b)