release_tool: Fix broken `-s` command for `other-components.yml`.
The logic dealing with the different yml files is complex, and
performing set operations on them is what ultimately caused `-s`
operation to not work for components in `other-components.yml`. Turns
out that we don't actually need the complex rules for all the tests to
pass (which I'm trusting for this work), we can just divide them
cleanly into docker and git groups.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>