test: gateway_and_one_device: User correct mender.conf on gateway
Previously, we were updating the mender.conf for the gateway device with
the values coming from the local device, iow setting ServerURL to the
gateway url. This actually works just fine, as now the client in the
gateway will use the gateway itself to talk to the backend...

Fixing for correctness of the test.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>