integ tests: Improve pytest exception hook for new fixtures
The newly developed fixtures for mender-gateway add gateway and the
local device(s) as separate MenderDevice (or MenderDeviceGroup)
instances. Modify the pytest hook to support more than one instance and
handle the list of devices or groups correctly.

Additionally, add mender-gateway to the list of services to print the
logs for and hard-code the mender-client one.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>