ci(scp): Fall back to the legacy SCP protocol
New versions of the scp program uses the SFTP protocol by
default. However SFTP is not installed in our qemu hosts, and hence
this does not work.

The fix is simple. Fall back to the legacy functionality. The SCP
protocol, though aged, works just fine for our needs.

This section was missed when doing 251bb499ef54542.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>