chore: enhance wait_for_connect making it more robust
TestAccessEnterprise.test_upgrades occasionally fails because
mender-connect running on the device is not connected when performing
troubleshoot add-on tests. This can happen because the tenant upgrade
triggers a workflow that resets the JWT tokens, and it can take a few
seconds for the Mender client to authorize again.

This change makes sure wait_for_connect waits for two consecutive checks
with 5 seconds interval before declaring the mender-connect client
component is available.

Changelog: None
Ticket: None

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>