test: increase recv timeout to 5 seconds
When running multiple tests in parallel, there is a high risk the shell
running in the Qemu device and controlled by mender-connect is not fast
enough at sending the output to deviceconnect. We increase the timeout
from one second to five to make sure we receive the full output before
proceeding with the tests.

Changelog: None
Ticket: None

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>