test(test_monitorclient_alert_email): Improve robustness
Disable the checks of each scenario when the verification is done.

This test covers different (independent!) scenarios, creating and
enabling checks for each of them. In some cases the alerts from one
scenario can leak to the next, and then the logic of "expect the alert
in the last received message" is too naive.

The last scenario (expiration) needs now to create the check from
scratch because it is disabled previously.

Moved also the mender-monitor paths definitions to a single place.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>